rename: 파일 경로 변경 및 불필요한 파일 삭제
