+++ src/components/CheckBox/CheckBox.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { CheckedBox, NoneCheckedBox } from '../asset';
+import { CheckedBox, NoneCheckedBox } from '../../asset';
 
 const CheckBoxImg = styled.img`
   width: 2.4rem;
