+++ src/components/Header/Header.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled, { css } from 'styled-components';
-import { LogoIcon, BackArrowIcon } from '../asset';
+import { LogoIcon, BackArrowIcon } from '../../asset';
 import { useNavigate } from 'react-router-dom';
 
 const HeaderContainer = styled.div`
